--- src/Console/Commands/ImportSourceFactory.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Matchish\ScoutElasticSearch\Console\Commands;
-
-interface ImportSourceFactory
-{
-    public static function from(string $className): ImportSource;
-}
