+++ src/Jobs/Stages/CreateWriteIndex.php
@@ -3,10 +3,10 @@
 namespace Matchish\ScoutElasticSearch\Jobs\Stages;
 
 use Elasticsearch\Client;
+use Matchish\ScoutElasticSearch\Searchable\ImportSource;
 use Matchish\ScoutElasticSearch\ElasticSearch\Index;
 use Matchish\ScoutElasticSearch\ElasticSearch\WriteAlias;
 use Matchish\ScoutElasticSearch\ElasticSearch\DefaultAlias;
-use Matchish\ScoutElasticSearch\Console\Commands\ImportSource;
 use Matchish\ScoutElasticSearch\ElasticSearch\Params\Indices\Create;
 
 /**
