+++ src/Searchable/ImportSource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Matchish\ScoutElasticSearch\Console\Commands;
+namespace Matchish\ScoutElasticSearch\Searchable;
 
 interface ImportSource
 {
