+++ src/Searchable/ImportSourceFactory.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Matchish\ScoutElasticSearch\Searchable;
+
+interface ImportSourceFactory
+{
+    public static function from(string $className, array $scopes = []): ImportSource;
+}
