+++ tests/Integration/Jobs/Stages/PullFromSourceTest.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Tests\Integration\Jobs\Stages;
 
+use Matchish\ScoutElasticSearch\Searchable\DefaultImportSourceFactory;
 use stdClass;
 use App\Product;
 use Tests\IntegrationTestCase;
 use Matchish\ScoutElasticSearch\Jobs\Stages\PullFromSource;
-use Matchish\ScoutElasticSearch\Console\Commands\DefaultImportSourceFactory;
 
 final class PullFromSourceTest extends IntegrationTestCase
 {
