+++ tests/Unit/Jobs/ImportStagesTest.php
@@ -3,6 +3,7 @@
 namespace Tests\Unit\Jobs;
 
 use App\Product;
+use Matchish\ScoutElasticSearch\Searchable\DefaultImportSourceFactory;
 use Tests\TestCase;
 use Matchish\ScoutElasticSearch\Jobs\ImportStages;
 use Matchish\ScoutElasticSearch\Jobs\Stages\CleanUp;
@@ -10,7 +11,6 @@ use Matchish\ScoutElasticSearch\Jobs\Stages\RefreshIndex;
 use Matchish\ScoutElasticSearch\Jobs\Stages\PullFromSource;
 use Matchish\ScoutElasticSearch\Jobs\Stages\CreateWriteIndex;
 use Matchish\ScoutElasticSearch\Jobs\Stages\SwitchToNewAndRemoveOldIndex;
-use Matchish\ScoutElasticSearch\Console\Commands\DefaultImportSourceFactory;
 
 class ImportStagesTest extends TestCase
 {
